API/requirements.txt
google-genai
flask
dotenv
flask_cors
firebase_admin